(*) API 호출 방식 변경
   12개씩 나눠서 1시간마다 1번씩 API를 호출하고 파싱하는 방식은 API 호출 횟수도 많고 속도도 느려서 한번에 호출하고 내부에서 파싱하는 방식으로 변경

+++ app/src/main/java/com/example/mnemonic/weather/model/WeatherFormattedData.kt
@@ -7,7 +7,7 @@ data class WeatherFormattedDataPerHour(
     var humidity: Double? = 0.0,
     var snowfallAmount: Int? = 0,
     var weatherType: WeatherType? = WeatherType.Sunny,
-    var temperature: Double? = 0.0,
+    var temperature: Int? = 0,
     var temperatureLow: Double? = 0.0,  // no use
     var temperatureHigh: Double? = 0.0, // no use
     var eastWestWindSpeed: Double? = 0.0,
