+++ app/src/main/java/com/example/mnemonic/weather/model/WeatherFormattedDataPerDay.kt
@@ -2,6 +2,6 @@ package com.example.mnemonic.weather.model
 
 data class WeatherFormattedDataPerDay (
     val weatherFormattedDataPerHourList: MutableList<WeatherFormattedDataPerHour> = mutableListOf(),
-    var maximumTemperature: Double = Double.MIN_VALUE,
-    var minimumTemperature: Double = Double.MAX_VALUE
+    var maximumTemperature: Int = Int.MIN_VALUE,
+    var minimumTemperature: Int = Int.MAX_VALUE
 )
